Add SyncIO instances, relocate IO implementation
Including the old IO implementations for OutWatch, Handler, etc in
the outwatch.dom package was causing too many conflicts if one wanted
to specify a different effect type, since one could not import dom._ without
also bringing in the IO versions. So this commit will break user code, but all
users have to do is add:
```scala
import outwatch.dom.io._   // For Handler, OutWatch.render*, etc.
import outwatch.util.io._  // For Store
```
