+++ outwatch/src/main/scala/outwatch/dom/AsValueObservable.scala
@@ -28,4 +28,3 @@ object AsValueObservable extends AsValueObservableInstances0  {
     }
   }
 }
-
