+++ outwatch/src/main/scala/outwatch/dom/ManagedSubscriptions.scala
@@ -1,6 +1,6 @@
 package outwatch.dom
 
-import cats.effect.IO
+import cats.effect.SyncIO
 import cats.implicits._
 import monix.execution.{Cancelable, Scheduler}
 import monix.execution.cancelables.CompositeCancelable
@@ -9,8 +9,8 @@ import org.scalajs.dom
 trait ManagedSubscriptions {
 
   @inline def managed(subscription: () => Cancelable): VDomModifier = managedElement(_ => subscription())
-  @inline def managed(subscription: IO[Cancelable]): VDomModifier = managed(() => subscription.unsafeRunSync())
-  def managed(sub1: IO[Cancelable], sub2: IO[Cancelable], subscriptions: IO[Cancelable]*): VDomModifier = {
+  @inline def managed(subscription: SyncIO[Cancelable]): VDomModifier = managed(() => subscription.unsafeRunSync())
+  def managed(sub1: SyncIO[Cancelable], sub2: SyncIO[Cancelable], subscriptions: SyncIO[Cancelable]*): VDomModifier = {
     val composite = (sub1 :: sub2 :: subscriptions.toList).sequence.map(subs => CompositeCancelable(subs: _*))
     managed(composite)
   }
@@ -18,7 +18,7 @@ trait ManagedSubscriptions {
   def managedAction(action: Scheduler => Cancelable): VDomModifier = SchedulerAction(scheduler => managed(() => action(scheduler)))
 
   object managedElement {
-    def apply(subscription: dom.Element => Cancelable): VDomModifier = IO {
+    def apply(subscription: dom.Element => Cancelable): VDomModifier = SyncIO {
       var cancelable: Cancelable = null
       VDomModifier(
         dsl.onDomMount foreach { elem => cancelable = subscription(elem) },
@@ -26,7 +26,7 @@ trait ManagedSubscriptions {
       )
     }
 
-    def asHtml(subscription: dom.html.Element => Cancelable): VDomModifier = IO {
+    def asHtml(subscription: dom.html.Element => Cancelable): VDomModifier = SyncIO {
       var cancelable: Cancelable = null
       VDomModifier(
         dsl.onDomMount.asHtml foreach { elem => cancelable = subscription(elem) },
@@ -34,7 +34,7 @@ trait ManagedSubscriptions {
       )
     }
 
-    def asSvg(subscription: dom.svg.Element => Cancelable): VDomModifier = IO {
+    def asSvg(subscription: dom.svg.Element => Cancelable): VDomModifier = SyncIO {
       var cancelable: Cancelable = null
       VDomModifier(
         dsl.onDomMount.asSvg foreach { elem => cancelable = subscription(elem) },
