+++ outwatch/src/main/scala/outwatch/dom/helpers/EmitterBuilder.scala
@@ -2,7 +2,7 @@ package outwatch
 package dom.helpers
 
 import cats.{Monoid, Functor, Bifunctor}
-import cats.effect.IO
+import cats.effect.SyncIO
 import monix.execution.{Cancelable, Scheduler}
 import monix.reactive.{Observable, Observer, OverflowStrategy}
 import org.scalajs.dom.{Element, Event, html, svg}
@@ -104,7 +104,7 @@ object EmitterBuilder {
 
   implicit class ModifierActions[O](val builder: EmitterBuilder[O, VDomModifier]) extends AnyVal {
     def withLatest[T](emitter: EmitterBuilder[T, VDomModifier]): EmitterBuilder[(O, T), VDomModifier] = new CustomEmitterBuilder[(O, T), VDomModifier]({ sink =>
-      IO {
+      SyncIO {
         var lastValue: js.UndefOr[T] = js.undefined
         VDomModifier(
           emitter foreach { lastValue = _ },
