+++ outwatch/src/test/scala/outwatch/StoreSpec.scala
